# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PharmaCare Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: paymentMode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale (POS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithItems"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithItems"

  # ─── SALE RETURNS ────────────────────────────────────────────────────────────
  /sale-returns:
    get:
      operationId: listSaleReturns
      tags: [saleReturns]
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sale returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleReturn"
    post:
      operationId: createSaleReturn
      tags: [saleReturns]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleReturnBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleReturn"
  /sale-returns/{id}:
    get:
      operationId: getSaleReturn
      tags: [saleReturns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale return with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleReturnWithItems"

  # ─── MISSED SALES ────────────────────────────────────────────────────────────
...
  /stock-audits:
    get:
      operationId: listStockAudits
      tags: [stockAudits]
      responses:
        "200":
          description: List of stock audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockAudit"
    post:
      operationId: createStockAudit
      tags: [stockAudits]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockAuditBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAuditWithItems"
  /stock-audits/{id}:
    get:
      operationId: getStockAudit
      tags: [stockAudits]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock audit with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAuditWithItems"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNo:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        date:
          type: string
        subtotal:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        changeAmount:
          type: number
        paymentMode:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, invoiceNo, date, subtotal, discountAmount, totalAmount, paidAmount, changeAmount, paymentMode, createdAt]
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        batchId:
          type: ["integer", "null"]
        batchNo:
          type: ["string", "null"]
        saleUnit:
          type: string
        quantity:
          type: number
        unitQuantity:
          type: number
        salePrice:
          type: number
        discountPercent:
          type: number
        totalAmount:
          type: number
        isControlled:
          type: boolean
        prescriptionNote:
          type: ["string", "null"]
      required: [id, medicineId, medicineName, saleUnit, quantity, unitQuantity, salePrice, discountPercent, totalAmount, isControlled]
    SaleWithItems:
      allOf:
        - $ref: "#/components/schemas/Sale"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SaleItem"
          required: [items]
    CreateSaleBody:
      type: object
      properties:
        customerId:
          type: ["integer", "null"]
        date:
          type: string
        discountAmount:
          type: number
        paidAmount:
          type: number
        paymentMode:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateSaleItemBody"
      required: [date, paidAmount, paymentMode, items]
    CreateSaleItemBody:
      type: object
      properties:
        medicineId:
          type: integer
        batchId:
          type: ["integer", "null"]
        saleUnit:
          type: string
        quantity:
          type: number
        salePrice:
          type: number
        discountPercent:
          type: number
        prescriptionNote:
          type: ["string", "null"]
      required: [medicineId, saleUnit, quantity, salePrice, discountPercent]

    # ─── SALE RETURNS ──────────────────────────────────────────────────────────
    SaleReturn:
      type: object
      properties:
        id:
          type: integer
        saleId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        date:
          type: string
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, date, totalAmount, createdAt]
    SaleReturnItem:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        batchId:
          type: ["integer", "null"]
        quantity:
          type: number
        salePrice:
          type: number
        totalAmount:
          type: number
      required: [id, medicineId, medicineName, quantity, salePrice, totalAmount]
    SaleReturnWithItems:
      allOf:
        - $ref: "#/components/schemas/SaleReturn"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SaleReturnItem"
          required: [items]
    CreateSaleReturnBody:
      type: object
      properties:
        saleId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        date:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateSaleReturnItemBody"
      required: [date, items]
    CreateSaleReturnItemBody:
      type: object
      properties:
        medicineId:
          type: integer
        batchId:
          type: ["integer", "null"]
        quantity:
          type: number
        salePrice:
          type: number
      required: [medicineId, quantity, salePrice]

    # ─── MISSED SALES ──────────────────────────────────────────────────────────
...
    StockAudit:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, date, status, createdAt]
    StockAuditItem:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        systemUnits:
          type: integer
        physicalUnits:
          type: integer
        physicalPacks:
          type: number
        varianceUnits:
          type: integer
      required: [id, medicineId, medicineName, systemUnits, physicalUnits, physicalPacks, varianceUnits]
    StockAuditWithItems:
      allOf:
        - $ref: "#/components/schemas/StockAudit"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/StockAuditItem"
          required: [items]
    CreateStockAuditBody:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateStockAuditItemBody"
      required: [title, date, items]
    CreateStockAuditItemBody:
      type: object
      properties:
        medicineId:
          type: integer
        physicalUnits:
          type: integer
        physicalPacks:
          type: number
      required: [medicineId, physicalUnits, physicalPacks]

    # ─── SETTINGS ──────────────────────────────────────────────────────────────
...